Changed background image for the dashboard (#71)
* Added a new model for 'Reports' to submit officer committee reports and
communications to be added to the agenda. Added a form where brothers
could submit these reports and a way for brothers to view their reports,
edit and delete them. Added a way for the secretary to view all of the
reports, delete them or delete all of them.

* Fixed minor bugs with agenda submission. Added ALL positions to
positions list

* Changed background of dashboard

* Fixed migrations