Updated attendance and excuse tracking (#74)
* Started finishing up the code for skipping MABs

* Finished the MAB exclusion thing

* Fixed attendance to only track brothers who were active when the event
was created. Created methods to check attendance ratios that can be
called on in template rather than passing these as context. Made events
use multi-table inheritance rather than abstract class inheritance,
meaning excuses can now be made for all events, not just chapter events

* Added a way for chairs and leadership to add brothers to the list of
eligible_attendees after the fact. When brothers are added to committees
they're automatically added to committee events as eligible_attendees if
the commitee event has not occurred yet. If they're removed from a
committee they are taken off of the list of eligible_attendees for
future committee meetings.

* Made the committee assingment function more efficient

* added comments

* Changed one of the datetime.dates to datetime.datetime

* Added an excuse form to other events and added better UI for resolving
and reading excuses for the secretary. Additionally, the recruitment
chair will be able to read and address recruitment events as well.

* Finished up everything related to excuses, brothers can now write and
track excuses for all events (with the exception of study_table events
and committee meetings). Additionally, brothers can easily manage and
track their attendance for each type of event through a fraction at the
bottom of each panel on the brother page.

* Added comments to multiple views and methods, generalized some redundant
views

* Cleaned up somoe views and forms

* Update dashboard/forms.py

Co-authored-by: Jared Cassarly <jaredcassarly@case.edu>

* Addressed some minor issues

* Added documentation and resolved redundancies

Co-authored-by: Jared Cassarly <jaredcassarly@case.edu>